update ui of event type create/edit

+++ churchinfo/EditEventTypes.php
@@ -15,9 +15,9 @@
  *
  *
  *  Modified by Stephen Shaffer, Oct 2006
- *  feature changes - added recurring defaults and customizable attendance count 
+ *  feature changes - added recurring defaults and customizable attendance count
  *  fields
- *       
+ *
  ******************************************************************************/
 require "Include/Config.php";
 require "Include/Functions.php";
@@ -35,11 +35,6 @@ if ( answer )
     window.location="EventEditor.php?Opp=" + Opp + "&Action=delete"
 }
 </script>
-<table width="100%" align="center" cellpadding="4" cellspacing="0">
-  <tr>
-    <td align="center"><input type="button" class="btn" value="<?= gettext("Back to Menu") ?>" Name="Exit" onclick="javascript:document.location='Menu.php';"></td>
-  </tr>
-</table>
 <?php
 //
 //  process the ACTION button inputs from the form page
@@ -47,39 +42,39 @@ if ( answer )
 $editing='FALSE';
 $tyid = $_POST["EN_tyid"];
 
-switch ($_POST['Action']){
-case "DELETE":
-  $ctid = $_POST['EN_ctid'];
+if(strpos($_POST['Action'], 'DELETE_', 0) === 0) {
+  $ctid = substr($_POST['Action'], 7);
   $sSQL = "DELETE FROM eventcountnames_evctnm WHERE evctnm_countid='$ctid' LIMIT 1";
   RunQuery($sSQL);
-  break;
-  
-case "ADD":
-  $newCTName = $_POST["newCountName"];
-  $theID=$_POST["EN_tyid"];  
-  $sSQL = "INSERT eventcountnames_evctnm (evctnm_eventtypeid, evctnm_countname) VALUES ('$theID','$newCTName')";
-  RunQuery($sSQL);
-  break;
+} else {
+  switch ($_POST['Action']){
+  case "ADD":
+    $newCTName = $_POST["newCountName"];
+    $theID=$_POST["EN_tyid"];
+    $sSQL = "INSERT eventcountnames_evctnm (evctnm_eventtypeid, evctnm_countname) VALUES ('$theID','$newCTName')";
+    RunQuery($sSQL);
+    break;
 
-case "Save Name":
-  $editing='FALSE';
-  $eName = $_POST["newEvtName"];
-  $theID=$_POST["EN_tyid"];
-  $sSQL = "UPDATE event_types SET type_name='$eName' WHERE type_id='$theID'";
-  RunQuery($sSQL);
-  $theID='';
-  $_POST['Action']='';
-  break;
-  
-case "Save Time":
-  $editing='FALSE';
-  $eTime = $_POST["newEvtStartTime"];
-  $theID=$_POST["EN_tyid"];  
-  $sSQL = "UPDATE event_types SET type_defstarttime='$eTime' WHERE type_id='$theID'";
-  RunQuery($sSQL);
-  $theID='';
-  $_POST['Action']='';
-  break;
+  case "NAME":
+    $editing='FALSE';
+    $eName = $_POST["newEvtName"];
+    $theID=$_POST["EN_tyid"];
+    $sSQL = "UPDATE event_types SET type_name='$eName' WHERE type_id='$theID'";
+    RunQuery($sSQL);
+    $theID='';
+    $_POST['Action']='';
+    break;
+
+  case "TIME":
+    $editing='FALSE';
+    $eTime = $_POST["newEvtStartTime"];
+    $theID=$_POST["EN_tyid"];
+    $sSQL = "UPDATE event_types SET type_defstarttime='$eTime' WHERE type_id='$theID'";
+    RunQuery($sSQL);
+    $theID='';
+    $_POST['Action']='';
+    break;
+  }
 }
 
 // Get data for the form as it now exists.
@@ -88,7 +83,7 @@ $rsOpps = RunQuery($sSQL);
 $aRow = mysql_fetch_array($rsOpps, MYSQL_BOTH);
 extract($aRow);
 $aTypeID = $type_id;
-$aTypeName = $type_name;             
+$aTypeName = $type_name;
 $aDefStartTime = $type_defstarttime;
     $aStartTimeTokens = explode(":", $aDefStartTime);
     $aEventStartHour = $aStartTimeTokens[0];
@@ -98,19 +93,19 @@ $aDefRecurDOM = $type_defrecurDOM;
 $aDefRecurDOY = $type_defrecurDOY;
 $aDefRecurType = $type_defrecurtype;
 switch ($aDefRecurType){
-    case "none": 
+    case "none":
        $recur="None";
        break;
-    case "weekly": 
+    case "weekly":
        $recur="Weekly on ".$aDefRecurDOW;
        break;
-    case "monthly": 
+    case "monthly":
        $recur="Monthly on ".date('dS',mktime(0,0,0,1,$aDefRecurDOM,2000));
        break;
-    case "yearly": 
+    case "yearly":
        $recur="Yearly on ".substr($aDefRecurDOY,5);
        break;
-    default: 
+    default:
        $recur="None";
 }
 
@@ -127,89 +122,83 @@ if($numCounts)
         $cRow = mysql_fetch_array($cOpps, MYSQL_BOTH);
         extract($cRow);
         $cCountID[$c] = $evctnm_countid;
-        $cCountName[$c] = $evctnm_countname;     
+        $cCountName[$c] = $evctnm_countname;
      }
 
-}                 
+}
 
 // Construct the form
 ?>
-<table width="100%" align="center" cellpadding="1" cellspacing="0">
-<form method="POST" action="EditEventTypes.php" name="EventTypeEditForm">
+<div class='box'>
+  <div class='box-header'>
+    <h3 class='box-title'><?= gettext("Edit Event Type") ?></h3>
+  </div>
+
+  <form method="POST" action="EditEventTypes.php" name="EventTypeEditForm">
   <input type="hidden" name="EN_tyid" value="<?= $aTypeID ?>">
-  <caption>
-    <h3><?= gettext("Edit Event Type") ?></h3>
-  </caption>
+  <input type="hidden" name="EN_ctid" value="<?= $cCountID[$c] ?>">
+
+<table class='table'>
   <tr>
-     <td class="LabelColumn" width="15%">
-        <strong><?= gettext("Event Type") . ":" . $aTypeID ?></strong>
-     </td>
-     <td class="TextColumn" colspan="3" width="85%">
-        <input type="text" name="newEvtName" value="<?= $aTypeName ?>" size="30" maxlength="35"> <input type="submit" Class="SmallText" Name="Action" value="<?= gettext("Save Name") ?>" class="btn">
-              <script language="javascript"> 
-                document.UpdateEventNames.newEvtName.focus() 
-              </script>
-          </form>
-     </td>
+    <td class="LabelColumn" width="15%">
+      <strong><?= gettext("Event Type") . ":" . $aTypeID ?></strong>
+    </td>
+    <td class="TextColumn" width="35%">
+      <input type="text" class="form-control" name="newEvtName" value="<?= $aTypeName ?>" size="30" maxlength="35" autofocus />
+    </td>
+    <td class="TextColumn" width="50%">
+      <button type="submit" Name="Action" value="NAME" class="btn btn-default"><?= gettext("Save Name") ?></button>
+    </td>
   </tr>
   <tr>
-   <td class="LabelColumn" width="15%">
+    <td class="LabelColumn" width="15%">
       <strong><?= gettext("Recurrance Pattern") ?></strong>
-   </td>
-   <td class="TextColumn" width="35%">
+    </td>
+    <td class="TextColumn" width="35%">
       <?= $recur ?>
-   </td>
-   <td class="LabelColumn" align="left" width="15%">
-        <strong><?= gettext("Start Time") ?></strong>
-   </td>
-   <td class="TextColumn" width="35%">
-        <form method="POST" action="EditEventTypes.php" name="EventTypeEditForm">  
-        <input type="hidden" name="EN_tyid" value="<?= $aTypeID ?>"> 
-        <input type="hidden" name="Action" value="Save Time">    
-        <select name="newEvtStartTime" size="1" onchange="javascript:this.form.submit()">
-         <?php createTimeDropdown(7,18,15,$aEventStartHour,$aEventStartMins); ?>
-        </select>
-        &nbsp;<span class="SmallText"><?= gettext("[format: HH:MM]") ?></span>
-        </form>
-   </td>
-   </tr>
+    </td>
+    <td class="TextColumn" width="50%">
+      <select class='form-control' name="newEvtStartTime" size="1" onchange="javascript:$('#newEvtStartTimeSubmit').click()">
+        <?php createTimeDropdown(7,18,15,$aEventStartHour,$aEventStartMins); ?>
+      </select>
+      <button class='hidden' type="submit" name="Action" value="TIME" id="newEvtStartTimeSubmit"></button>
+    </td>
+  </tr>
 
-  <tr><td colspan="4"></td></tr>
    <tr>
-      <td class="LabelColumn" align="center" width="15%" rowspan="<?= $nr ?>" colspan="1">
+      <td class="LabelColumn" width="15%" rowspan="<?= $nr ?>" colspan="1">
         <strong><?= gettext("Attendance Counts") ?></strong>
       </td>
     </tr>
     <?php
-    for($c = 1; $c <=$numCounts; $c++){ 
+    for($c = 1; $c <=$numCounts; $c++){
       ?>
       <tr>
-        <td class="TextColumn" colspan="1" width="35%"><?= $cCountName[$c] ?></td>
-        <td class="TextColumn" colspan="2" width="50%">
-           <form name="DelEvCount" action="EditEventTypes.php" method="POST">
-           <input type="hidden" name="EN_ctid" value="<?= $cCountID[$c] ?>">
-           <input type="hidden" name="EN_tyid" value="<?= $aTypeID ?>">
-           <input type="submit" class="SmallText" name="Action" value="<?= gettext("DELETE") ?>" class="btn")">
-           </form>
+        <td class="TextColumn" width="35%"><?= $cCountName[$c] ?></td>
+        <td class="TextColumn" width="50%">
+          <button type="submit" name="Action" value="DELETE_<?=  $cCountID[$c] ?>" class="btn btn-default"><?= gettext("Remove") ?></button>
         </td>
       </tr>
-     <?php 
-     }     
+     <?php
+     }
      ?>
       <tr>
-        <td class="TextColumn" colspan="1" width="35%">
-           <form name="AddEvCount" action="EditEventTypes.php" method="POST">          
-           <input type="text" name="newCountName" length="20" value="New Attendance Count">
+        <td class="TextColumn" width="35%">
+           <input class='form-control' type="text" name="newCountName" length="20" placeholder="New Attendance Count" />
         </td>
-        <td class="TextColumn" colspan="2" width="50%">
-
-           <input type="hidden" name="EN_ctid" value="new">
-           <input type="hidden" name="EN_tyid" value="<?= $aTypeID ?>">
-           <input type="submit" class="SmallText" name="Action" value="<?= gettext("ADD") ?>" class="btn")">
-           </form>
+        <td class="TextColumn" width="50%">
+           <button type="submit" name="Action" value="ADD" class="btn btn-default"><?= gettext("Add counter") ?></button>
         </td>
       </tr>
-
 </table>
+</form>
+</div>
+
+<div>
+  <a href="EventNames.php" class='btn btn-default'>
+    <i class='glyphicon glyphicon-chevron-left'></i>
+    <?= gettext("Return to Event Types") ?>
+  </a>
+</div>
 
 <?php require "Include/Footer.php" ?>
